- Added Vue and VueSocket.io - Basic refactor for routes
