+++ bower.json
@@ -21,6 +21,8 @@
   "dependencies": {
     "jquery": "^3.1.0",
     "dropzone": "https://github.com/enyo/dropzone.git#^4.3.0",
-    "ion-sound": "ionsound#^3.0.7"
+    "ion-sound": "ionsound#^3.0.7",
+    "vue": "^2.5.16",
+    "bootstrap4": "^4.1.1"
   }
 }
