+++ package.json
@@ -11,7 +11,8 @@
     "mime": "^1.3.4",
     "multer": "^1.1.0",
     "socket.io": "^1.4.8",
-    "string.prototype.startswith": "^0.2.0"
+    "string.prototype.startswith": "^0.2.0",
+    "vue": "^2.5.16",
+    "vue-socket.io": "^2.1.1-b"
   }
 }
-
